Refactor / resolve propertyFlag tests
